fix(web): resolve lint errors in product editor (#97)
